+++ editor/src/liquidator/asset/gltf/MaterialStep.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "GLTFImportData.h"
+
+namespace liquidator {
+
+/**
+ * @brief Load materials into registry
+ *
+ * @param importData GLTF import data
+ */
+void loadMaterials(GLTFImportData &importData);
+
+} // namespace liquidator
