+++ editor/src/liquidator/asset/gltf/MeshStep.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "GLTFImportData.h"
+
+namespace liquidator {
+
+/**
+ * @brief Load meshes into registry
+ *
+ * @param importData GLTF import data
+ */
+void loadMeshes(GLTFImportData &importData);
+
+} // namespace liquidator
