+++ editor/src/liquidator/asset/gltf/PrefabStep.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "GLTFImportData.h"
+
+namespace liquidator {
+
+/**
+ * @brief Load prefabs into asset registry
+ *
+ * @param importData GLTF import data
+ */
+void loadPrefabs(GLTFImportData &importData);
+
+} // namespace liquidator
