+++ editor/src/liquidator/asset/gltf/SkeletonStep.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "GLTFImportData.h"
+
+namespace liquidator {
+
+/**
+ * @brief Load skeletons into registry
+ *
+ * @param importData GLTF import data
+ */
+void loadSkeletons(GLTFImportData &importData);
+
+} // namespace liquidator
