+++ editor/src/liquidator/asset/gltf/TextureStep.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "GLTFImportData.h"
+
+namespace liquidator {
+
+/**
+ * @brief Load textures into registry
+ *
+ * @param importData GLTF import data
+ */
+void loadTextures(GLTFImportData &importData);
+
+} // namespace liquidator
