+++ editor/src/liquidator/asset/gltf/TinyGLTF.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#define TINYGLTF_NO_INCLUDE_STB_IMAGE
+#define TINYGLTF_NO_INCLUDE_JSON
+#define TINYGLTF_NO_STB_IMAGE_WRITE
+#define TINYGLTF_USE_CPP14
+
+#include <json/json.hpp>
+#include <stb/stb_image.h>
+#include <tinygltf/tiny_gltf.h>
