+++ editor/src/liquidator/asset/gltf/TinyGLTFImpl.cpp
@@ -1,0 +1,4 @@
+#include "liquid/core/Base.h"
+
+#define TINYGLTF_IMPLEMENTATION
+#include "TinyGLTF.h"
